refactor: move WithUnixSocketListener to a linux specific file
[ upstream commit 067f790d76f0e2bd1ce4e84f7f3795c0365eeebc ]
